test(repo): isolate and extend IAM connection coverage
